// packages/wp-react-lib/src/types.ts
export interface DgSettings {
    react_ui_url:          string;
    react_api_url:         string | null;
    apache_superset_url:   string | boolean | null;
    react_search_type:     string;
    react_menu_type:       string;
    languages:             Languages;
    landing_page_url:      string;
    google_analytics_code: string;
    name:                  string;
    description:           string;
    site_logo:             number;
    site_icon:             number;
}

export interface Languages {
    [key: string]: Language;
}

export interface Language {
    enable:          number;
    locale:          string;
    name:            string;
    translation:     string;
    date:            string;
    time:            string;
    flag:            string;
    wpseo_og_locale: string;
}